chore: re-organize message modules, fix icon size, convo reset properly rebuilds Tree
